+++ source_gen/source_gen/lib/src/builder.dart
@@ -37,12 +37,16 @@ class _Builder extends Builder {
   final Map<String, List<String>> buildExtensions;
 
   /// Wrap [_generators] to form a [Builder]-compatible API.
-  _Builder(this._generators,
+  _Builder(
+      this._generators,
       {String formatOutput(String code),
       String generatedExtension: '.g.dart',
       List<String> additionalOutputExtensions: const [],
       bool isStandalone: false,
-      bool requireLibraryDirective: true,
+      @Deprecated(
+          'Library directives are no longer required for part generation. '
+          'This option will be removed in v0.8.0.')
+          bool requireLibraryDirective: false,
       String header})
       : _generatedExtension = generatedExtension,
         buildExtensions = {
@@ -50,6 +54,7 @@ class _Builder extends Builder {
         },
         _isStandalone = isStandalone,
         formatOutput = formatOutput ?? _formatter.format,
+        // ignore: deprecated_member_use
         _requireLibraryDirective = requireLibraryDirective,
         _header = header ?? defaultFileHeader {
     if (_generatedExtension == null) {
@@ -167,23 +172,30 @@ class PartBuilder extends _Builder {
   /// If `null`, the content of [defaultFileHeader] is used.
   /// If [header] is an empty `String` no header is added.
   ///
-  /// May set [requireLibraryDirective] to `false` in order to opt-in to
-  /// supporting a `2.0.0-dev` feature of `part of` being usable without an
-  /// explicit `library` directive. Developers should restrict their `pubspec`
+  /// May set [requireLibraryDirective] to `true` in order to opt-out of the
+  /// Dart `2.0.0-dev` feature of `part of` being usable without an explicit
+  /// `library` directive. Developers should restrict their `pubspec`
   /// accordingly:
   /// ```yaml
   /// sdk: '>=2.0.0-dev <2.0.0'
   /// ```
-  PartBuilder(List<Generator> generators,
+  ///
+  /// This option will be removed in version 0.8.0 of `source_gen`.
+  PartBuilder(
+      List<Generator> generators,
       {String formatOutput(String code),
       String generatedExtension: '.g.dart',
       List<String> additionalOutputExtensions: const [],
-      bool requireLibraryDirective: true,
+      @Deprecated(
+          'Library directives are no longer required for part generation. '
+          'This option will be removed in v0.8.0.')
+          bool requireLibraryDirective: false,
       String header})
       : super(generators,
             formatOutput: formatOutput,
             generatedExtension: generatedExtension,
             additionalOutputExtensions: additionalOutputExtensions,
+            // ignore: deprecated_member_use
             requireLibraryDirective: requireLibraryDirective,
             header: header);
 }